+++ CHANGES.md
@@ -4,7 +4,7 @@ Added the io.aviso.toolchest.vars namespace.
 
 Added the io.aviso.toolchest.collections/update-if? and into+ functions.
 
-Added the io.aviso.toolchest.metadata/copying-meta macro.
+Added the io.aviso.toolchest.metadata/merging-meta macro.
 
 ## 0.1.3 - 24 Dec 2015